+++ src/de/dhbw/shake_it_app/data/operator/DataOperator.java
@@ -2,7 +2,6 @@ package de.dhbw.shake_it_app.data.operator;
 
 import de.dhbw.shake_it_app.data.DataProvider;
 import de.dhbw.shake_it_app.data.model.Session;
-import de.dhbw.shake_it_app.data.model.User;
 
 public class DataOperator {
 	
@@ -19,7 +18,7 @@ public class DataOperator {
 		return dataOperator;
 	}
 	
-	public int returnCurrLocationIndex(int clubID){
+	public int returnCurrLocationIndex(long clubID){
 		double cumulatedValues = 0;
 		int amountValues = 0;
 		Session[] sessions = (Session[]) DataProvider.get().getModel(DataProvider.Session, "filter=locationID&value="+clubID);
@@ -32,7 +31,7 @@ public class DataOperator {
 		return (int) Math.round(cumulatedValues/amountValues);
 	}
 	
-	public int returnOverallLocationIndex(int clubID){
+	public int returnOverallLocationIndex(long clubID){
 		double cumulatedValues = 0;
 		int amountValues = 0;
 		Session[] sessions = (Session[]) DataProvider.get().getModel(DataProvider.Session, "filter=locationID&value="+clubID);
@@ -45,7 +44,7 @@ public class DataOperator {
 		return (int) Math.round(cumulatedValues/amountValues);
 	}
 	
-	public int returnOverallUserIndex(int userID){
+	public int returnOverallUserIndex(long userID){
 		double cumulatedValues = 0;
 		int amountValues = 0;
 		Session[] sessions = (Session[]) DataProvider.get().getModel(DataProvider.Session, "filter=userID&value="+userID);
